+++ src/routes/index.tsx
@@ -17,9 +17,9 @@ import InstitutionUsers from "@/pages/dashboard/institution/users";
 import InstitutionProfile from "@/pages/dashboard/institution/profile";
 import InstitutionAccount from "@/pages/dashboard/institution/account";
 import Transactions from "@/pages/dashboard/institution/account/transactions";
-import MyRequest from "@/pages/dashboard/requests/pages/MyRequest";
-import ReviewRequest from "@/pages/dashboard/requests/pages/ReviewRequest";
-import ManageRequestPage from "@/pages/dashboard/requests/pages/ManageRequestPage";
+import MyRequest from "@/pages/dashboard/requests/pages/admin/MyRequest";
+import ReviewRequest from "@/pages/dashboard/requests/pages/admin/ReviewRequest";
+import ManageRequestPage from "@/pages/dashboard/requests/pages/admin/ManageRequestPage";
 import Templates from "@/pages/dashboard/templates";
 import ViewTransactions from "@/pages/dashboard/pricing/view-transactions";
 import ModifyRequest from "@/pages/dashboard/requests/edit-requests";
